dashboard/requirements.txt
streamlit==1.32.0
psycopg2-binary==2.9.9
pandas==2.2.1
plotly==5.20.0
pyvis==0.3.2
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.4.1
scipy==1.12.0
